refactor feature directory, removed subdirectories
