+++ src/features/spectrumDataSlice.js
